refactor: calendar view types moved to a separate file
